Chore: Move notifications/notify to services
WATER-1791

Moves any notification sending to the NotificationsApiClient classes.
