+++ src/shared/lib/connectors/services/water/NotificationsApiClient.js
@@ -1,6 +1,6 @@
 const { APIClient } = require('@envage/hapi-pg-rest-api');
 const urlJoin = require('url-join');
-const { http } = require('@envage/water-abstraction-helpers');
+const { http, serviceRequest } = require('@envage/water-abstraction-helpers');
 
 const getEndpoint = serviceUrl => urlJoin(serviceUrl, 'notification');
 
@@ -33,6 +33,19 @@ class NotificationsApiClient extends APIClient {
     const pagination = { page: 1, perPage: 1 };
     return this.findMany(filter, sort, pagination);
   };
+
+  async sendNotifyMessage (messageRef, recipient, personalisation) {
+    const url = urlJoin(this.config.serviceUrl, 'notify', messageRef);
+    const body = { recipient, personalisation };
+
+    try {
+      const response = await serviceRequest.post(url, { body });
+      return response.body;
+    } catch (err) {
+      this.logger.error('Error sending notify message', { error: err });
+      return err;
+    };
+  }
 };
 
 module.exports = NotificationsApiClient;
